Moved location features into separate file
